+++ app/workspace/departments/[slug]/page.tsx
@@ -78,7 +78,7 @@ export default async function DepartmentLaunchpadPage({
   }
 
   const [
-    { departmentAgents, externalAgents, myAgents },
+    { departmentAgents, externalGroups, myAgents },
     canManageTemplates,
     collapsedPrefResult,
   ] = await Promise.all([
@@ -142,7 +142,7 @@ export default async function DepartmentLaunchpadPage({
 
       <DepartmentLaunchpadContent
         departmentAgents={departmentAgents}
-        externalAgents={externalAgents}
+        externalGroups={externalGroups}
         myAgents={myAgents}
         departmentSlug={department.slug}
         canManageTemplates={canManageTemplates}